devex: moved testable logic from fix-race-condition-served-in-drafts into a helpers file and adjusted the imports accordingly
