+++ scripts/dynamo/fix-race-condition-served-in-drafts.helpers.ts
@@ -3,48 +3,8 @@
 import { Case } from '@shared/business/entities/cases/Case';
 import { DocketEntry } from '@shared/business/entities/DocketEntry';
 import { DynamoDBClient, GetItemCommand } from '@aws-sdk/client-dynamodb';
-import {
-  type ScriptConfig,
-  parseArgsAndEnvVars,
-} from '../helpers/parseArgsAndEnvVars';
 import { aggregatePartiesForService } from '@shared/business/utilities/aggregatePartiesForService';
-import { createApplicationContext } from '@web-api/applicationContext';
-import { readFileSync } from 'fs';
 import { unmarshall } from '@aws-sdk/util-dynamodb';
-
-const scriptConfig: ScriptConfig = {
-  description:
-    'fix-race-condition-served-in-drafts - Resolves issue where a document that had been ' +
-    'served returned to the drafts folder.',
-  environment: { dynamoDbTableName: 'DYNAMODB_TABLE_NAME' },
-  parameters: {
-    docketEntryId: {
-      position: 1,
-      required: true,
-      type: 'string',
-    },
-    docketNumber: {
-      position: 0,
-      required: true,
-      type: 'string',
-    },
-    pathToJsonRequest: {
-      position: 3,
-      required: true,
-      type: 'string',
-    },
-    performUpdate: {
-      default: false,
-      type: 'boolean',
-    },
-    timestamp: {
-      position: 2,
-      required: true,
-      type: 'string',
-    },
-  },
-  requireActiveAwsSession: true,
-};
 
 export const getDocumentFromDynamo = async ({
   docketEntryId,
@@ -194,32 +154,3 @@ export const fixRaceConditionServedInDrafts = async (
     });
   }
 };
-
-// eslint-disable-next-line @typescript-eslint/no-floating-promises
-(async () => {
-  const {
-    docketEntryId,
-    docketNumber,
-    pathToJsonRequest,
-    performUpdate,
-    timestamp,
-  } = parseArgsAndEnvVars(scriptConfig) as {
-    docketEntryId: string;
-    docketNumber: string;
-    pathToJsonRequest: string;
-    performUpdate: boolean;
-    timestamp: string;
-  };
-
-  const request = JSON.parse(readFileSync(pathToJsonRequest, 'utf-8'));
-
-  const applicationContext = createApplicationContext({});
-
-  await fixRaceConditionServedInDrafts(applicationContext, {
-    docketEntryId,
-    docketNumber,
-    performUpdate,
-    request,
-    timestamp,
-  });
-})();
